backend/requirements.txt
Django==3.2.3
djangorestframework==3.15.1
djoser==2.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pillow==10.3.0
gunicorn==22.0.0